+++ src/remote_store/__init__.py
@@ -18,6 +18,7 @@ from remote_store._path import RemotePath
 from remote_store._registry import Registry, register_backend
 from remote_store._store import Store
 from remote_store.ext.batch import BatchResult, batch_copy, batch_delete, batch_exists
+from remote_store.ext.glob import glob_files
 from remote_store.ext.transfer import download, transfer, upload
 
 __version__ = "0.10.0"
@@ -53,6 +54,8 @@ __all__ = [
     "batch_delete",
     "batch_copy",
     "batch_exists",
+    # Glob operations
+    "glob_files",
     # Transfer operations
     "upload",
     "download",